requirements.txt
Flask-Bootstrap
opencv-python
numpy
flask
pillow
gunicorn
requests